EasyPeasy1.1

+++ src/controller/ItemListaController.java
@@ -13,16 +13,24 @@ import java.util.Scanner;
 import static view.Menus.menuItem;
 import static view.Listas.exibirListaDeDesejosById;
 
+/*
+Controla o CRUD de itens dentro de uma lista de desejos.
+Usado por idosos (criar/editar) e familiares (acompanhar status).
+*/
 public class ItemListaController {
+    static Scanner scn = new Scanner(System.in);
 
-    public static void editarItem(int idlista, Scanner scn, itemDAO itemDao, ListaDeDesejosDAO listaDao) {
+    /*
+    Edita um item existente (nome, descrição, quantidade, loja, link).
+     */
+    public static void editarItem(int idlista, itemDAO itemDao, ListaDeDesejosDAO listaDao) {
         try {
-            exibirListaDeDesejosById(idlista, listaDao, itemDao);
+            exibirListaDeDesejosById(idlista, listaDao, itemDao);//exibe lista e seuse itens para lista para o usuário logado escolher (metodo da view -> Listas)
             System.out.printf("Digite o id do item que deseja editar: ");
             int idItem = scn.nextInt();
             scn.nextLine();
 
-            itemDao.update(idItem);
+            itemDao.update(idItem);//chama menu de atualização do DAO
 
         } catch (SQLException e) {
             System.out.printf("ERRO! Não foi possível editar este item!");
@@ -30,7 +38,10 @@ public class ItemListaController {
         }
     }
 
-    public static void cadastrarItens(int idlista, itemDAO itemDao, Scanner scn) throws SQLException {
+    /*
+    Cadastra um novo item e cria histórico PENDENTE automaticamente
+     */
+    public static void cadastrarItens(int idlista, itemDAO itemDao) throws SQLException {
         System.out.println("Digite [0] para voltar");
         System.out.println("-----------CriandoItem-----------");
         System.out.printf("Nome: ");
@@ -74,17 +85,20 @@ public class ItemListaController {
             return;
         }
 
-        item i = new item(idlista, nomeItem, descricaoItem, quant, nomeLoja, link);
-        itemDao.save(i);
+        item i = new item(idlista, nomeItem, descricaoItem, quant, nomeLoja, link);// Cria objeto item
+        itemDao.save(i); // Salva no banco → retorna ID gerado
     }
 
-    public static void deletarItem(int idlista, Scanner scn, itemDAO itemDao, ListaDeDesejosDAO listaDao) {
+    /*
+    Deleta um item com confirmação Y/N (dentro do DAO).
+     */
+    public static void deletarItem(int idlista, itemDAO itemDao, ListaDeDesejosDAO listaDao) {
 
         try {
             while (true) {
 
                 try {
-                    exibirListaDeDesejosById(idlista, listaDao, itemDao);
+                    exibirListaDeDesejosById(idlista, listaDao, itemDao);//exibe lista e seuse itens para lista para o usuário logado escolher (metodo da view -> Listas)
                     System.out.println("Digite [0] para voltar");
                     System.out.printf("Digite o id do item que deseja excluir: ");
                     int idItem = scn.nextInt();
@@ -95,7 +109,7 @@ public class ItemListaController {
                         return;
                     }
 
-                    itemDao.delete(idItem);
+                    itemDao.delete(idItem);// metodo de confirmação Y/N dentro do DAO
 
                 } catch (InputMismatchException e) {
                     System.out.println("ERRO! Digite apenas números");
@@ -108,8 +122,11 @@ public class ItemListaController {
         }
     }
 
+    /*
+    Tela principal de gerenciamento de itens.
+     */
     public static void telaItens(int idlista, Connection conn) throws SQLException {
-        Scanner scn = new Scanner(System.in);
+        //DAOs instanciadas para buscas e acesso de metodos
         UsuarioDAO usuarioDao = new UsuarioDAO(conn);
         ListaDeDesejosDAO listaDao = new ListaDeDesejosDAO(conn);
         itemDAO itemDao = new itemDAO(conn);
@@ -126,11 +143,11 @@ public class ItemListaController {
 
                 switch (opcao) {
                     case 1 -> {
-                        cadastrarItens(idlista, itemDao, scn);
+                        cadastrarItens(idlista, itemDao); // Cadastra item e cria histórico (metodo de ListaItemController)
                         return;
                     }
-                    case 2 -> editarItem(idlista, scn, itemDao, listaDao);
-                    case 3 -> deletarItem(idlista, scn, itemDao, listaDao);
+                    case 2 -> editarItem(idlista, itemDao, listaDao); // Edita item (metodo de ListaItemController)
+                    case 3 -> deletarItem(idlista, itemDao, listaDao); // Deleta item (metodo de ListaItemController)
                     case 0 -> {
                         System.out.println("Voltando...");
                         return;
